emulator: thread enable mask doesn't need to be a 64 bit value

+++ tools/emulator/core.c
@@ -82,7 +82,7 @@ struct Core
 	Thread *threads;
 	struct Breakpoint *breakpoints;
 	int singleStepping;
-	uint64_t threadEnableMask;
+	uint32_t threadEnableMask;
 	int halt;
 	int stopOnFault;
 	int enableTracing;
